[core] use `/var/run/dd-agent` for temp files
It was already implemented for PidFile by default, added it to pickle
files, as it should fix all issues with `dd-agent info` reporting
incorrectly. It doesn't break anything, as it checks first to see if
`/var/run/dd-agent` exists, and fallback to the current behaviour if
not.

Also:
- move PidFile and Platform from util.py to their own files in `utils/`
- as a consequence, move `checks/utils.py` into `utils/tailfile.py`
