+++ agent.py
@@ -40,9 +40,9 @@ from util import (
     EC2,
     get_hostname,
     get_os,
-    PidFile,
     Watchdog,
 )
+from utils.pidfile import PidFile
 from utils.flare import configcheck, Flare
 
 # Constants
