+++ tests/test_cassandra_jmx.py
@@ -4,7 +4,7 @@ from nose.plugins.attrib import attr
 import threading
 from aggregator import MetricsAggregator
 from dogstatsd import Server
-from util import PidFile
+from utils.pidfile import PidFile
 import os
 from jmxfetch import JMXFetch
