+++ tests/test_java_jmx.py
@@ -10,7 +10,7 @@ from mock import patch
 # datadog
 from aggregator import MetricsAggregator
 from dogstatsd import Server
-from util import PidFile
+from utils.pidfile import PidFile
 from jmxfetch import JMXFetch
 from common import AgentCheckTest
